move to other file and memoize status as well
