+++ benchexec/tablegenerator/react-table/src/components/Table/index.js
@@ -1,0 +1,12 @@
+// This file is part of BenchExec, a framework for reliable benchmarking:
+// https://github.com/sosy-lab/benchexec
+//
+// SPDX-FileCopyrightText: 2019-2020 Dirk Beyer <https://www.sosy-lab.org>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+import FilterInputField from "./FilterInputField";
+import MinMaxFilterInputField from "./MinMaxFilterInputField";
+import StatusFilter from "./StatusFilter";
+
+export { FilterInputField, MinMaxFilterInputField, StatusFilter };
